fix modulewidget deletion crash
do i know why it was crashing? no i do not. but it looks like we should
try to do any rack poking within the initial construction on the render
thread. i tried to shift the deletion of moduleWidget into the action
queue (on the render thread), but putting it in a lambda also caused a
crash? idfk